@@ -14,15 +14,19 @@
         }
     ],
     "require": {
-        "php": "^7.2",
+        "php": "^7.4 || ^8.0",
         "ext-json": "*"
     },
     "require-dev": {
         "eventjet/coding-standard": "^3.0",
-        "infection/infection": "^0.15.0",
+        "infection/infection": "^0.20.0",
         "maglnet/composer-require-checker": "^2.0",
-        "phpstan/phpstan": "^0.12.3",
-        "phpunit/phpunit": "^8.0"
+        "phpstan/extension-installer": "^1.0",
+        "phpstan/phpstan": "^0.12.32",
+        "phpstan/phpstan-phpunit": "^0.12.11",
+        "phpunit/phpunit": "^9.4",
+        "psalm/plugin-phpunit": "^0.13.0",
+        "vimeo/psalm": "^4.0"
     },
     "config": {
         "sort-packages": true
@@ -44,15 +48,17 @@
             "@check-deps",
             "@cs-check",
             "@phpstan",
+            "@psalm",
             "@phpunit",
             "@infection"
         ],
         "check-deps": "vendor/bin/composer-require-checker",
         "cs-check": "vendor/bin/phpcs",
         "cs-fix": "vendor/bin/phpcbf",
-        "infection": "vendor/bin/infection --threads=4 --min-msi=90 --min-covered-msi=90",
+        "infection": "vendor/bin/infection --threads=4",
         "infection-xdebug": "@composer run infection -- --initial-tests-php-options='-d zend_extension=xdebug.so'",
-        "phpstan": "vendor/bin/phpstan analyse src/ --level 8",
-        "phpunit": "vendor/bin/phpunit"
+        "phpstan": "vendor/bin/phpstan analyse",
+        "phpunit": "vendor/bin/phpunit",
+        "psalm": "vendor/bin/psalm"
     }
 }
